Pavement and Dirt settings
These dont work correctly at the moment due to the queue buffering multiple inputs which is less noticeable when going up but left and right send you into the side of the screen and ignore the code made to stop that exact issue. With any luck Mike will be able to help with the queue issue and then hedges, Dirt and Pavement code will all fall into place. Water and logs work fine as is, as does the rail section. The only issue left are the cars since bunner needs more states to help avoid being run over